Started work on comment filtering
Comments ought to be approved by an admin before being posted to the
general public, as they don't require any user registration. Spam or
vandalism comments need a way to be removed. However, for a user, I want
them to see their comment on the site right away. I'll do this by
storing a cookie on their machine for each comment they make, that way I
can add logic to show that comment only for them, but not for anyone
else, until approved by the admin